+++ fs/rsext4/src/ext4/mod.rs
@@ -32,10 +32,9 @@ mod sync;
 mod system_zone;
 
 pub use fs::{Ext4FileSystem, FileSystemStats};
-pub use lookup::{file_entry_exisr, file_entry_exist, find_file};
+pub use lookup::{file_entry_exist, find_file};
 pub use mkfs::{
-    BlcokGroupLayout, BlockGroupLayout, FsLayoutInfo, MkfsOptions, compute_fs_layout, mkfs,
-    mkfs_with_options,
+    BlockGroupLayout, FsLayoutInfo, MkfsOptions, compute_fs_layout, mkfs, mkfs_with_options,
 };
 pub use mount::{MountOptions, mount, mount_with_options, mount_with_options_and_observer};
 pub use owned::{